requirements.txt
autopep8==1.6.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.5
pycodestyle==2.8.0
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
toml==0.10.2
Werkzeug==2.0.2
